+++ src/domains/Musica/services/serviceMusica.ts
@@ -1,9 +1,14 @@
 import prisma from '../../../../config/prismaClient.js';
 import { Prisma } from '@prisma/client';
+import { InvalidParamError, QueryError } from '../../../../errors/index.js';
 
 class ServiceMusica {
     // Cria nova música
     async criarMusica(body: Prisma.MusicaCreateInput) {
+
+        if (body.nome == null){
+            throw new InvalidParamError("Nome não informado!")
+        }
         const musica: Prisma.MusicaCreateInput = {
             nome: body.nome,
             genero: body.genero,
@@ -21,6 +26,17 @@ class ServiceMusica {
 
     // Deleta música com id especifico
     async deletarMusica(id: string) {
+        const musicaEncontrada = await prisma.musica.findUnique({
+            where:{
+                id: id
+            }
+        })
+        if (musicaEncontrada == null){
+            throw new QueryError("Id inválido!")
+        }
+        if (id == null){
+            throw new InvalidParamError("Id não informado!")
+        }
         try {
             await prisma.musica.delete({
                 where: {
@@ -55,6 +71,17 @@ class ServiceMusica {
 
     // Retoma a música com id especificado
     async listarMusicaID(id: string) {
+        const musicaEncontrada = await prisma.musica.findUnique({
+            where:{
+                id: id
+            }
+        })
+        if (musicaEncontrada == null){
+            throw new QueryError("Id inválido!")
+        }
+        if (id == null){
+            throw new InvalidParamError("Id não informado!")
+        }
         try {
             const musica = await prisma.musica.findUnique({
                 where: {
@@ -77,6 +104,17 @@ class ServiceMusica {
     // Atualiza informações da música de id especificado
     async atualizaMusica(id: string, body: Prisma.MusicaUpdateInput) {
         const musicaUpdate: Prisma.MusicaUpdateInput = { ...body };
+        const musicaEncontrada = await prisma.musica.findUnique({
+            where:{
+                id: id
+            }
+        })
+        if (musicaEncontrada == null){
+            throw new QueryError("Id inválido!")
+        }
+        if (id == null){
+            throw new InvalidParamError("Id não informado!")
+        }
         try {
             const musica = await prisma.musica.update({
                 where: {
@@ -92,6 +130,28 @@ class ServiceMusica {
 
     // Vincula um artista à musica
     async vinculaMusicaArtista(artistaId: string, musicaId: string) {
+        const musicaEncontrada = await prisma.musica.findUnique({
+            where:{
+                id: musicaId
+            }
+        })
+        const artistaEncontrado = await prisma.artista.findUnique({
+            where:{
+                id: artistaId
+            }
+        })
+        if (musicaEncontrada == null){
+            throw new QueryError("Id da música inválido!")
+        }
+        if (artistaEncontrado == null){
+            throw new QueryError("Id do artista inválido!")
+        }
+        if (musicaId == null){
+            throw new InvalidParamError("Id da música não informado!")
+        }
+        if (artistaId == null){
+            throw new InvalidParamError("Id do artista não informado!")
+        }
         try {
             const autoria = await prisma.autoria.create({
                 data: {
@@ -115,6 +175,28 @@ class ServiceMusica {
 
     // Desvincula um artista à musica
     async desvinculaMusicaArtista(where: Prisma.AutoriaWhereUniqueInput) {
+        const musicaEncontrada = await prisma.musica.findUnique({
+            where:{
+                id: where.artistaId_musicaId?.musicaId
+            }
+        })
+        const artistaEncontrado = await prisma.artista.findUnique({
+            where:{
+                id: where.artistaId_musicaId?.artistaId
+            }
+        })
+        if (musicaEncontrada == null){
+            throw new QueryError("Id da música inválido!")
+        }
+        if (artistaEncontrado == null){
+            throw new QueryError("Id do artista inválido!")
+        }
+        if ( where.artistaId_musicaId?.musicaId == null){
+            throw new InvalidParamError("Id da música não informado!")
+        }
+        if (where.artistaId_musicaId?.artistaId == null){
+            throw new InvalidParamError("Id do artista não informado!")
+        }
         try {
             const autoria = await prisma.autoria.delete({
                 where: where,
